requirements.txt
requests~=2.32.5
fastapi~=0.116.0
uvicorn[standard]~=0.35.0
pdfplumber~=0.11.7

openai~=1.100.1
httpx~=0.28.1
qdrant-client~=1.15.1
langchain~=0.3.26
torch~=2.8.0
transformers~=4.55.2
langchain-openai~=0.3.30
langchain-core~=0.3.74
langchain-huggingface==0.3.1
langchain-ollama==0.3.6
langchain-qdrant==0.2.0
langgraph~=0.6.5
async-timeout~=5.0.1
PyYAML~=6.0.2
FlagEmbedding~=1.3.5
sentence-transformers~=5.1.0
python-multipart~=0.0.20

kiwipiepy~=0.21.0
pydantic~=2.11.7